+++ R/first_differences.R
@@ -22,11 +22,9 @@
 #' trend("additive")        # additive preprocessing
 #' trend("differences", 1)  # order 1 first differences
 #' trend("differences", -1) # order of first differences automatically estimated
-trend <- function(type = "additive", n = -1, transform_features = FALSE) {
-  if (! (is.character(type) && length(type) == 1))
-    stop("type argument in trend_prepro should be a character")
-  if (! type %in% c("none", "additive", "multiplicative", "differences"))
-    stop("type argument in trend_prepro should be none, additive, multiplicative or differences")
+trend <- function(type = c("additive", "multiplicative", "differences", "none"), 
+                  n = -1, transform_features = TRUE) {
+  type <- match.arg(type)
   if (! (is.numeric(n) && length(n) == 1 && n >= -1 && floor(n) == n))
     stop("n parameter should be an integer scalar value >= -1")
   if (! (is.logical(transform_features) && length(transform_features) == 1))